+++ source/bxdatatools/src/configuration/ui/variant_repository_viewer.cc
@@ -259,7 +259,8 @@ namespace datatools {
         if (!_repository_) {
           return;
         }
-
+        _read_only_ = !_repository_->get_ui_config().writable_at_start;
+        
         // Top:
         _repository_name_display_label_->setAlignment(Qt::AlignLeft);
         _repository_name_display_label_->setTextFormat(Qt::PlainText);